chore(auth): spilitted the code of auth component, data, and service
